fix: #1732 only re-insert text node if it's parent is text node too while deleting

+++ frontend/app_flowy/packages/appflowy_editor/lib/src/service/internal_key_event_handlers/backspace_handler.dart
@@ -120,7 +120,8 @@ KeyEventResult _backDeleteToPreviousTextNode(
 ) {
   if (textNode.next == null &&
       textNode.children.isEmpty &&
-      textNode.parent?.parent != null) {
+      textNode.parent?.parent != null &&
+      textNode.parent is TextNode) {
     transaction
       ..deleteNode(textNode)
       ..insertNode(textNode.parent!.path.next, textNode)
